add 'brick pack' command
